add NettyClientChannelManager to refactor AbstractRpcRemotingClient.
