+++ core/src/main/java/io/seata/core/rpc/netty/RmRpcClient.java
@@ -17,35 +17,27 @@ package io.seata.core.rpc.netty;
 
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelHandler.Sharable;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.handler.timeout.IdleStateEvent;
 import io.netty.util.concurrent.EventExecutorGroup;
 import io.seata.common.exception.FrameworkErrorCode;
 import io.seata.common.exception.FrameworkException;
 import io.seata.common.thread.NamedThreadFactory;
-import io.seata.common.util.NetUtil;
 import io.seata.core.model.Resource;
 import io.seata.core.model.ResourceManager;
 import io.seata.core.protocol.AbstractMessage;
-import io.seata.core.protocol.HeartbeatMessage;
 import io.seata.core.protocol.RegisterRMRequest;
 import io.seata.core.protocol.RegisterRMResponse;
 import io.seata.core.rpc.netty.NettyPoolKey.TransactionRole;
-import io.seata.core.rpc.netty.pool.ClientPoolKeyGenerator;
-import io.seata.core.rpc.netty.pool.RmClientPoolKeyGenerator;
-import org.apache.commons.pool.impl.GenericKeyedObjectPool.Config;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.function.Function;
 
 import static io.seata.common.Constants.DBKEYS_SPLIT_CHAR;
 
@@ -53,44 +45,25 @@ import static io.seata.common.Constants.DBKEYS_SPLIT_CHAR;
  * The type Rm rpc client.
  *
  * @author jimin.jm @alibaba-inc.com
+ * @author zhaojun
  * @date 2018 /10/10
  */
 @Sharable
 public final class RmRpcClient extends AbstractRpcRemotingClient {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(RmRpcClient.class);
-
     private ResourceManager resourceManager;
-
     private static volatile RmRpcClient instance;
-    private final ConcurrentMap<String, Object> channelLocks = new ConcurrentHashMap<>();
-    private final ConcurrentMap<String, NettyPoolKey> poolKeyMap = new ConcurrentHashMap<>();
-    private final ConcurrentMap<String, Channel> channels = new ConcurrentHashMap<>();
     private String customerKeys;
     private final AtomicBoolean initialized = new AtomicBoolean(false);
-    private static final int MAX_MERGE_SEND_THREAD = 1;
     private static final long KEEP_ALIVE_TIME = Integer.MAX_VALUE;
     private static final int MAX_QUEUE_SIZE = 20000;
-    private static final int SCHEDULE_INTERVAL_MILLS = 5;
-    private static final String MERGE_THREAD_PREFIX = "rpcMergeMessageSend";
-    private final NettyClientConfig rmClientConfig;
-    
-    private final ClientPoolKeyGenerator poolKeyGenerator;
+    private String applicationId;
+    private String transactionServiceGroup;
     
     private RmRpcClient(NettyClientConfig nettyClientConfig, EventExecutorGroup eventExecutorGroup,
                         ThreadPoolExecutor messageExecutor) {
         super(nettyClientConfig, eventExecutorGroup, messageExecutor, TransactionRole.RMROLE);
-        this.rmClientConfig = nettyClientConfig;
-        poolKeyGenerator = new RmClientPoolKeyGenerator(applicationId, transactionServiceGroup, resourceManager, customerKeys);
-    }
-
-    /**
-     * Sets resource manager.
-     *
-     * @param resourceManager the resource manager
-     */
-    public void setResourceManager(ResourceManager resourceManager) {
-        this.resourceManager = resourceManager;
     }
 
     /**
@@ -117,192 +90,128 @@ public final class RmRpcClient extends AbstractRpcRemotingClient {
             synchronized (RmRpcClient.class) {
                 if (null == instance) {
                     NettyClientConfig nettyClientConfig = new NettyClientConfig();
-                    final ThreadPoolExecutor threadPoolExecutor = new ThreadPoolExecutor(
+                    final ThreadPoolExecutor messageExecutor = new ThreadPoolExecutor(
                         nettyClientConfig.getClientWorkerThreads(), nettyClientConfig.getClientWorkerThreads(),
                         KEEP_ALIVE_TIME, TimeUnit.SECONDS,
                         new LinkedBlockingQueue<>(MAX_QUEUE_SIZE),
                         new NamedThreadFactory(nettyClientConfig.getRmDispatchThreadPrefix(),
                             nettyClientConfig.getClientWorkerThreads()),
                         new ThreadPoolExecutor.CallerRunsPolicy());
-                    instance = new RmRpcClient(nettyClientConfig, null, threadPoolExecutor);
+                    instance = new RmRpcClient(nettyClientConfig, null, messageExecutor);
                 }
             }
         }
         return instance;
     }
-
+    
+    /**
+     * Sets application id.
+     *
+     * @param applicationId the application id
+     */
+    public void setApplicationId(String applicationId) {
+        this.applicationId = applicationId;
+    }
+    
+    /**
+     * Sets transaction service group.
+     *
+     * @param transactionServiceGroup the transaction service group
+     */
+    public void setTransactionServiceGroup(String transactionServiceGroup) {
+        this.transactionServiceGroup = transactionServiceGroup;
+    }
+    
+    /**
+     * Sets resource manager.
+     *
+     * @param resourceManager the resource manager
+     */
+    public void setResourceManager(ResourceManager resourceManager) {
+        this.resourceManager = resourceManager;
+    }
+    
+    /**
+     * Gets customer keys.
+     *
+     * @return the customer keys
+     */
+    public String getCustomerKeys() {
+        return customerKeys;
+    }
+    
+    /**
+     * Sets customer keys.
+     *
+     * @param customerKeys the customer keys
+     */
+    public void setCustomerKeys(String customerKeys) {
+        this.customerKeys = customerKeys;
+    }
+    
     @Override
     public void init() {
         if (initialized.compareAndSet(false, true)) {
             super.init();
         }
     }
-
+    
     @Override
-    protected Config getNettyPoolConfig() {
-        Config poolConfig = new Config();
-        poolConfig.maxActive = rmClientConfig.getMaxPoolActive();
-        poolConfig.minIdle = rmClientConfig.getMinPoolIdle();
-        poolConfig.maxWait = rmClientConfig.getMaxAcquireConnMills();
-        poolConfig.testOnBorrow = rmClientConfig.isPoolTestBorrow();
-        poolConfig.testOnReturn = rmClientConfig.isPoolTestReturn();
-        poolConfig.lifo = rmClientConfig.isPoolLifo();
-        return poolConfig;
+    public void destroy() {
+        super.destroy();
+        initialized.getAndSet(false);
+        instance = null;
     }
-
+    
     @Override
-    protected NettyPoolKey.TransactionRole getTransactionRole() {
-        return TransactionRole.RMROLE;
+    protected Function<String, NettyPoolKey> getPoolKeyFunction() {
+        return (serverAddress) -> {
+            String resourceIds = customerKeys == null ? getMergedResourceKeys() : customerKeys;
+            if (LOGGER.isInfoEnabled()) {
+                LOGGER.info("RM will register :" + resourceIds);
+            }
+            RegisterRMRequest message = new RegisterRMRequest(applicationId, transactionServiceGroup);
+            message.setResourceIds(resourceIds);
+            return new NettyPoolKey(NettyPoolKey.TransactionRole.RMROLE, serverAddress, message);
+        };
     }
-
+    
     @Override
-    public void userEventTriggered(ChannelHandlerContext ctx, Object evt) {
-        if (evt instanceof IdleStateEvent) {
-            IdleStateEvent idleStateEvent = (IdleStateEvent)evt;
-            if (idleStateEvent == IdleStateEvent.READER_IDLE_STATE_EVENT) {
-                if (LOGGER.isInfoEnabled()) {
-                    LOGGER.info("RmRpcClient channel" + ctx.channel() + " idle.");
-                }
-                try {
-                    String serverAddress = NetUtil.toStringAddress(ctx.channel().remoteAddress());
-                    nettyClientKeyPool.invalidateObject(poolKeyMap.get(serverAddress), ctx.channel());
-                } catch (Exception exx) {
-                    LOGGER.error(exx.getMessage());
-                } finally {
-                    releaseChannel(ctx.channel(), getAddressFromContext(ctx));
-                }
-            }
-            if (idleStateEvent == IdleStateEvent.WRITER_IDLE_STATE_EVENT) {
-                try {
-                    sendRequest(ctx.channel(), HeartbeatMessage.PING);
-                } catch (Throwable throwable) {
-                    LOGGER.error("", "send request error", throwable);
-                }
-            }
-        }
+    protected String getTransactionServiceGroup() {
+        return transactionServiceGroup;
     }
-
+    
     @Override
-    protected void releaseChannel(Channel channel, String serverAddress) {
-        if (null == channel || null == serverAddress) { return; }
-        try {
-            synchronized (channelLocks.get(serverAddress)) {
-                Channel ch = channels.get(serverAddress);
-                if (null == ch) {
-                    nettyClientKeyPool.returnObject(poolKeyMap.get(serverAddress), channel);
-                    return;
-                }
-                if (ch.compareTo(channel) == 0) {
-                    if (LOGGER.isInfoEnabled()) {
-                        LOGGER.info("return to pool, rm channel:" + channel);
-                    }
-                    destroyChannel(serverAddress, channel);
-                } else {
-                    nettyClientKeyPool.returnObject(poolKeyMap.get(serverAddress), channel);
-                }
-            }
-        } catch (Exception exx) {
-            LOGGER.error(exx.getMessage());
-        }
-    }
-
-    @Override
-    protected Channel connect(String serverAddress) {
-        Channel channelToServer = channels.get(serverAddress);
-        if (channelToServer != null) {
-            channelToServer = getExistAliveChannel(channelToServer, serverAddress);
-            if (null != channelToServer) {
-                return channelToServer;
-            }
-        }
+    public void onRegisterMsgSuccess(String serverAddress, Channel channel, Object response,
+                                     AbstractMessage requestMessage) {
         if (LOGGER.isInfoEnabled()) {
-            LOGGER.info("will connect to " + serverAddress);
+            LOGGER.info(
+                "register RM success. server version:" + ((RegisterRMResponse)response).getVersion()
+                    + ",channel:" + channel);
         }
-        channelLocks.putIfAbsent(serverAddress, new Object());
-        synchronized (channelLocks.get(serverAddress)) {
-            return doConnect(serverAddress);
-        }
-    }
-
-    /**
-     * Connect channel.
-     *
-     * @param serverAddress the server address
-     * @return the channel
-     */
-    private Channel doConnect(String serverAddress) {
-        Channel channelToServer = channels.get(serverAddress);
-        if (channelToServer != null && channelToServer.isActive()) {
-            return channelToServer;
-        }
-        Channel channelFromPool;
-        try {
-            poolKeyMap.put(serverAddress, poolKeyGenerator.generate(serverAddress));
-            channelFromPool = nettyClientKeyPool.borrowObject(poolKeyMap.get(serverAddress));
-        } catch (Exception exx) {
-            LOGGER.error(FrameworkErrorCode.RegisterRM.getErrCode(), "register RM failed.", exx);
-            throw new FrameworkException("can not register RM,err:" + exx.getMessage());
-        }
-        return channelFromPool;
-    }
-
-    private Channel getExistAliveChannel(Channel rmChannel, String serverAddress) {
-        if (rmChannel.isActive()) {
-            return rmChannel;
-        } else {
-            int i = 0;
-            for (; i < NettyClientConfig.getMaxCheckAliveRetry(); i++) {
-                try {
-                    Thread.sleep(NettyClientConfig.getCheckAliveInternal());
-                } catch (InterruptedException exx) {
-                    LOGGER.error(exx.getMessage());
-                }
-                rmChannel = channels.get(serverAddress);
-                if (null == rmChannel || rmChannel.isActive()) {
-                    return rmChannel;
+        if (customerKeys == null) {
+            getClientChannelManager().registerChannel(serverAddress, channel);
+            String dbKey = getMergedResourceKeys();
+            RegisterRMRequest message = (RegisterRMRequest)requestMessage;
+            if (message.getResourceIds() != null) {
+                if (!message.getResourceIds().equals(dbKey)) {
+                    sendRegisterMessage(serverAddress, channel, dbKey);
                 }
             }
-            if (i == NettyClientConfig.getMaxCheckAliveRetry()) {
-                LOGGER.warn("channel " + rmChannel + " is not active after long wait, close it.");
-                releaseChannel(rmChannel, serverAddress);
-                return null;
-            }
         }
-        return null;
     }
 
     @Override
-    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
-        LOGGER.error(FrameworkErrorCode.ExceptionCaught.getErrCode(),
-            NetUtil.toStringAddress(ctx.channel().remoteAddress()) + "connect exception. " + cause.getMessage(),
-            cause);
-        releaseChannel(ctx.channel(), getAddressFromChannel(ctx.channel()));
-        if (LOGGER.isInfoEnabled()) {
-            LOGGER.info("remove exception rm channel:" + ctx.channel());
-        }
-        super.exceptionCaught(ctx, cause);
-    }
+    public void onRegisterMsgFail(String serverAddress, Channel channel, Object response,
+                                  AbstractMessage requestMessage) {
 
-    private void sendRegisterMessage(String serverAddress, Channel channel, String dbKey) {
-        RegisterRMRequest message = new RegisterRMRequest(applicationId, transactionServiceGroup);
-        message.setResourceIds(dbKey);
-        try {
-            super.sendAsyncRequestWithoutResponse(channel, message);
-        } catch (FrameworkException e) {
-            if (e.getErrcode() == FrameworkErrorCode.ChannelIsNotWritable
-                && serverAddress != null) {
-                releaseChannel(channel, serverAddress);
-                if (LOGGER.isInfoEnabled()) {
-                    LOGGER.info("remove channel:" + channel);
-                }
-            } else {
-                LOGGER.error("", "register failed", e);
-            }
-        } catch (TimeoutException e) {
-            LOGGER.error(e.getMessage());
+        if (response instanceof RegisterRMResponse && LOGGER.isInfoEnabled()) {
+            LOGGER.info(
+                "register RM failed. server version:" + ((RegisterRMResponse)response).getVersion());
         }
+        throw new FrameworkException("register RM failed.");
     }
-
+    
     /**
      * Register new db key.
      *
@@ -313,12 +222,12 @@ public final class RmRpcClient extends AbstractRpcRemotingClient {
         if (LOGGER.isInfoEnabled()) {
             LOGGER.info("register to RM resourceId:" + resourceId);
         }
-        if (channels.isEmpty()) {
-            reconnect(transactionServiceGroup);
+        if (getClientChannelManager().getChannels().isEmpty()) {
+            getClientChannelManager().reconnect(transactionServiceGroup);
             return;
         }
-        synchronized (channels) {
-            for (Map.Entry<String, Channel> entry : channels.entrySet()) {
+        synchronized (getClientChannelManager().getChannels()) {
+            for (Map.Entry<String, Channel> entry : getClientChannelManager().getChannels().entrySet()) {
                 String serverAddress = entry.getKey();
                 Channel rmChannel = entry.getValue();
                 if (LOGGER.isInfoEnabled()) {
@@ -328,64 +237,28 @@ public final class RmRpcClient extends AbstractRpcRemotingClient {
             }
         }
     }
-
-    @Override
-    public void sendResponse(long msgId, String serverAddress, Object msg) {
-        if (LOGGER.isInfoEnabled()) {
-            LOGGER.info("RmRpcClient sendResponse " + msg);
-        }
-        super.sendResponse(msgId, connect(serverAddress), msg);
-    }
-
-    /**
-     * Gets customer keys.
-     *
-     * @return the customer keys
-     */
-    public String getCustomerKeys() {
-        return customerKeys;
-    }
-
-    /**
-     * Sets customer keys.
-     *
-     * @param customerKeys the customer keys
-     */
-    public void setCustomerKeys(String customerKeys) {
-        this.customerKeys = customerKeys;
-    }
-
-    @Override
-    public void onRegisterMsgSuccess(String serverAddress, Channel channel, Object response,
-                                     AbstractMessage requestMessage) {
-
-        if (LOGGER.isInfoEnabled()) {
-            LOGGER.info(
-                "register RM success. server version:" + ((RegisterRMResponse)response).getVersion()
-                    + ",channel:" + channel);
-        }
-        if (customerKeys == null) {
-            synchronized (channels) {
-                channels.put(serverAddress, channel);
-            }
-            String dbKey = getMergedResourceKeys(resourceManager);
-            RegisterRMRequest message = (RegisterRMRequest)requestMessage;
-            if (message.getResourceIds() != null) {
-                if (!message.getResourceIds().equals(dbKey)) {
-                    sendRegisterMessage(serverAddress, channel, dbKey);
+    
+    private void sendRegisterMessage(String serverAddress, Channel channel, String dbKey) {
+        RegisterRMRequest message = new RegisterRMRequest(applicationId, transactionServiceGroup);
+        message.setResourceIds(dbKey);
+        try {
+            super.sendAsyncRequestWithoutResponse(channel, message);
+        } catch (FrameworkException e) {
+            if (e.getErrcode() == FrameworkErrorCode.ChannelIsNotWritable
+                && serverAddress != null) {
+                getClientChannelManager().releaseChannel(channel, serverAddress);
+                if (LOGGER.isInfoEnabled()) {
+                    LOGGER.info("remove channel:" + channel);
                 }
+            } else {
+                LOGGER.error("", "register failed", e);
             }
+        } catch (TimeoutException e) {
+            LOGGER.error(e.getMessage());
         }
     }
-
-    /**
-     * Gets merged resource keys.
-     *
-     * @param resourceManager the resource manager
-     * @return the merged resource keys
-     */
-    public String getMergedResourceKeys(ResourceManager resourceManager) {
-        //TODO
+    
+    private String getMergedResourceKeys() {
         Map<String, Resource> managedResources = resourceManager.getManagedResources();
         Set<String> resourceIds = managedResources.keySet();
         if (!resourceIds.isEmpty()) {
@@ -402,36 +275,5 @@ public final class RmRpcClient extends AbstractRpcRemotingClient {
             return sb.toString();
         }
         return null;
-    }
-
-    @Override
-    public void onRegisterMsgFail(String serverAddress, Channel channel, Object response,
-                                  AbstractMessage requestMessage) {
-
-        if (response instanceof RegisterRMResponse && LOGGER.isInfoEnabled()) {
-            LOGGER.info(
-                "register RM failed. server version:" + ((RegisterRMResponse)response).getVersion());
-        }
-        throw new FrameworkException("register RM failed.");
-    }
-
-    @Override
-    public void destroyChannel(String serverAddress, Channel channel) {
-        if (null == channel) { return; }
-        try {
-            if (channel.equals(channels.get(serverAddress))) {
-                channels.remove(serverAddress);
-            }
-            nettyClientKeyPool.returnObject(poolKeyMap.get(serverAddress), channel);
-        } catch (Exception exx) {
-            LOGGER.error("return channel to rmPool error:" + exx.getMessage());
-        }
-    }
-    
-    @Override
-    public void destroy() {
-        super.destroy();
-        initialized.getAndSet(false);
-        instance = null;
     }
 }
