--- core/src/main/java/io/seata/core/rpc/netty/pool/ClientPoolKeyGenerator.java
@@ -1,13 +1,0 @@
-package io.seata.core.rpc.netty.pool;
-
-import io.seata.core.rpc.netty.NettyPoolKey;
-
-/**
- * Client pool key generator.
- *
- * @author zhaojun
- */
-public interface ClientPoolKeyGenerator {
-    
-    NettyPoolKey generate(String serverAddress);
-}
